move comment to own type
