+++ client/src/types/Comment.d.ts
@@ -1,0 +1,10 @@
+interface Comment {
+  _id: string;
+  author: {
+    user_id: string;
+    email: string;
+  };
+  text: string;
+}
+
+export default Comment;
